requirements.txt
flask>=2.3.0,<3.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pyinstaller>=6.1.0,<7.0.0
werkzeug>=2.3.0,<3.0.0
jinja2>=3.1.0,<4.0.0
